test(ci/cd): yet another ci/cd test
This is just ridiculous. I pretty sure, that last 5 hours of my work
has been spent on trying to get my project to compile, because i'm using
a dependency, that uses a dependency, that uses a dependency, that uses
pkg-config, which is impossible to cross compile in a stream-lined way.
This has to be some kind of a joke, wtf.

Now I understand the need for web. Cross compiling is pure pain.