feat: remove windows build
This is riduculous. I have spent 4 days of my life trying to cross
compile a rust program.

All of that, because it's dependency is using a dependency which is
using a dependency that uses pkg-config, which requires setting env
varaibles to pass the build. This makes it impossible to use community
implemented solutions such as cross, thus making me download the sysroot
manually and try to compile to the target.

This is nuts.